Multiplicación directa hecho

+++ Proyecto-intermedio/funciones.h
@@ -1,2 +1,2 @@
-void multiplicacion_directa(const int  * A, const int  * B, int* C,const int N);
+void multiplicacion_directa(const double  * A, const double  * B, double* C,const int N);
